examples/calculatortest: Ensure we press the 8 button twice in test_division
Using find element by name twice to find the "8" button runs the risk of
finding the result text the second time since the text may have changed
to display 8 at that point. To avoid that, execute the find call once
and store the result so we can use it to call `click()` multiple times.