add xwayland support
previously the script was the process controller running all the stuff
below it. unfortunately for xwayland to work we need to run flask and
the actual test under kwin (so kwin can set the correct DISPLAY var). to
accomplish this some tricky steps had to be taken.
Kwin.with now either redirects the entire execution into a fork, or noop
yields.

we have two redirections now. USE_CUSTOM_BUS can redirect the Bus and
TEST_WITH_XWAYLAND redirects Kwin. This is getting a bit complicated ^^