implement key actions with level support
this is a bit awkward. synthesizes is dependent on the keycode but also
the level. e.g. when we get the key H on the selenium side that is the
level 1 (shift) + keycode for h actually, so when we then synthesize H
via kwin we first need to press the modifier, then the key, then release
again.

to support this, and future extensions for mouse input, the synthesizer
no longer takes the keycode as argument but rather the file path to a
json file with an array of actions. currently only keyboard actions are
supported and encompass the keycode and level so we can correctly
synthesize the relevant input.