Correct the CI dependency declarations for selenium-webdriver-at-spi.
The definition @same is reserved explicitly for projects within the same release unit to depend on each other.
Given that selenium-webdriver-at-spi lives outside of Plasma, it is improper and incorrect for @same to be used here to bring in KPipeWire and KWin.

The dependency definition also needs splitting between Qt 5 and Qt 6, as Plasma has opted to not support Qt 6 at all within the Plasma/5.27 series, and not support Qt 5 at all within the master branch series - yet we need to be able to support tests for both from the same branch.

This should resolve the issue reported at plasma/plasma-workspace#71

CCMAIL: plasma-devel@kde.org