inputsynth: rejigger level manipulation and layout management
previously the synth was at the mercy of the environment, now it should
be more resilient.

notable changes:

- key level is now determined on-the-fly by inspecting the key's
modifiers and resolving them to the relevant keys
- the relevant keys are identified by looking through all keys for
potential modifiers
- the keyboard layout is now fetched from org.kde.keyboard via dbus
- various other refinements

overall this should improve reliability of the synth. in particular with
regards to scenarios where the host keyboard layout is not 'us' and also
scenarios where the host has multiple layouts configured in the kcm e.g.
'at,us' and the nested kwin ended up using an unexpected layout.

the synth wire format has been changed to pass sequences of keys rather
than individual keys. this makes typing loads faster.
because of the format change event generation is now by default covering
an entire piece of text instead of a single key (single key is still
being used when atspi is used for synthesis)