inputsynth: implement proper actions processing (for keyboards)
this is an initial implementation for ActionChains processing. currently
this is limited to keyboards only. the wire format changed again to be
exactly that of the webdriver spec so we can just hand the data from
driver to inputsynth for the actions case. a fully featured
implementation of actionchains (including pointers and whatnot) needs to
be aware of tick concerns anyway, so all processing ideally should
happen in the inputsynth.

NOTE: for practical reasons I don't think this should ever be supported
outside nested kwins. we'd have to have two completely separate code
paths :((