implement better coordinate click support (wayland qt6 only)
this sits on top of a number of changes in the underlying infrastructure
to work around current shortcomings of atspi on wayland (specifically
the fact that the applications don't know their global coordinates and
as such we can't click based on accessible coordinates).

to work around this problem we have plasma-integration install an
extension object on the a11y bus that we can use to retrieve more
platform specific information about an accessible. specificaly we use it
to get the xdg foreign export handle as well as the accessible geometry.

using the export handle we are able to uniquely identify which window in
the compositor (kwin) is the one related to the accessible.

the geometry we can use to map between coordinate systems