Allow launch command override through env
Currently the application being tested is launched according to a
capability set in the testsuite script. This stops us from being able to
reuse the same test script for different packaging versions of the same
app in an automated context.

This patch adds handling for a `SELENIUM_OVERRIDE_LAUNCH=<command>`
environment variable which would let us alter the app launch at runtime
by overriding the `app` capability of the test script while keeping the
test content.

This means it would now be possible to only write a single test script
and let a CI run it for different versions of the app, for example a
snap-packaged version.