allow passing recording file by argument
this in particular makes it a bit easier to access the feature as part
of cmake code. previously one would have had to set an environment on
the test target which required a lot of syntax

the new argument gets stripped from ARGV so the actual test invocation
doesn't get tripped up by the foreign arg. this is a system we might
want to generalize at some point to control more stuff by arguments
rather than env vars